Refactored players code, split into submodules
