+++ src/picast/__main__.py
@@ -33,7 +33,9 @@ gi.require_version("GstVideo", "1.0")  # noqa: E402 # isort:skip
 gi.require_version("GdkX11", "3.0")  # noqa: E402 # isort:skip
 
 from .rtspsink import RtspSink  # noqa: E402 # isort:skip
-from .player import GstPlayer, VlcPlayer, NopPlayer  # noqa: E402 # isort:skip
+from .players.gst import GstPlayer  # noqa: E402 # isort:skip
+from .players.vlc import VlcPlayer  # noqa: E402 # isort:skip
+from .players.nop import NopPlayer  # noqa: E402 # isort:skip
 from .settings import Settings  # noqa: E402 # isort:skip
 from .wifip2p import WifiP2PServer  # noqa: E402 # isort:skip
